Implement changes as requested by Darren
